corrected test functions location
